feat(block-producer): address mempool review comments
