+++ crates/block-producer/src/block_builder/prover/block_witness.rs
@@ -7,7 +7,7 @@ use miden_objects::{
     notes::Nullifier,
     transaction::TransactionId,
     vm::{AdviceInputs, StackInputs},
-    BlockHeader, Digest, Felt, BLOCK_NOTE_TREE_DEPTH, ZERO,
+    BlockHeader, Digest, Felt, BLOCK_NOTE_TREE_DEPTH, MAX_BATCHES_PER_BLOCK, ZERO,
 };
 
 use crate::{
@@ -35,6 +35,9 @@ impl BlockWitness {
         mut block_inputs: BlockInputs,
         batches: &[TransactionBatch],
     ) -> Result<(Self, Vec<BlockAccountUpdate>), BuildBlockError> {
+        // This limit should be enforced by the mempool.
+        assert!(batches.len() <= MAX_BATCHES_PER_BLOCK);
+
         Self::validate_nullifiers(&block_inputs, batches)?;
 
         let batch_created_notes_roots = batches
